Moved methods up from base review to baseproposal. Removed redunant code.
