+++ proposals/tests/proposal_submission_tests.py
@@ -67,6 +67,25 @@ class BaseProposalTestCase(TestCase):
         receive a cached value."""
         self.proposal.refresh_from_db()
 
+    def check_subject_lines(self, outbox):
+        """
+        Make sure every outgoing email contains a reference number and the
+        text FETC-GW
+        """
+        for message in outbox:
+            subject = message.subject
+            self.assertTrue("FETC-GW" in subject)
+            self.assertTrue(self.proposal.reference_number in subject)
+
+    def set_relation_to_phd_student(self, supervisor: User):
+        """Phd student status means a supervisor is needed.
+        While the previous postdoc relation does not need a supervisor"""
+        self.proposal.relation = Relation.objects.get(
+            description_nl="als AIO / promovendus verbonden"
+        )
+        self.proposal.supervisor = supervisor
+        self.proposal.save()
+
 
 class ProposalSubmitTestCase(
     BaseViewTestCase,
@@ -160,9 +179,7 @@ class ProposalSubmitTestCase(
         """
         # Select the PHD relation, which needs a supervisor
         # but doesn't check for a study/course
-        self.proposal.relation = Relation.objects.get(pk=4)
-        self.proposal.supervisor = self.supervisor
-        self.proposal.save()
+        self.set_relation_to_phd_student(self.supervisor)
 
         self.proposal.study_set.all().delete()
         # Sanity checks to start
